Moving the xarch HWIntrtinsicInfo type into its own header file.
